Separates the menu and media pane from the VideoPane.
